useSearchParams使ってるところ　suspendで囲う  pageで宣言してる場合はコンポーネント化してpageを新たに作成して対処
