+++ my-app/src/app/work-log/daily/[date]/page.tsx
@@ -19,6 +19,7 @@ const CircleGraph = dynamic(() => import("./circle-graph/CircleGraph"), {
 });
 import MemoList from "./memo-list/MemoList";
 import dynamic from "next/dynamic";
+import { Suspense } from "react";
 
 /**
  * 日付詳細ページ
@@ -27,24 +28,26 @@ export default function DailyDetailPage() {
   const { selectedItemId, selectedItemTaskId, handleSelectItem } =
     DailyDetailPageParams();
   return (
-    <Stack direction="row" spacing={1} pt={3} px={2}>
-      {/**　左半分(メニュー/タスク) */}
-      <Stack width="50%" justifyContent={"space-between"}>
-        {/** メニュー */}
-        <DailyDetailMenu />
-        {/** タスク */}
-        <TaskList
-          selectedItemId={selectedItemId}
-          handleClickRow={handleSelectItem}
-        />
+    <Suspense>
+      <Stack direction="row" spacing={1} pt={3} px={2}>
+        {/**　左半分(メニュー/タスク) */}
+        <Stack width="50%" justifyContent={"space-between"}>
+          {/** メニュー */}
+          <DailyDetailMenu />
+          {/** タスク */}
+          <TaskList
+            selectedItemId={selectedItemId}
+            handleClickRow={handleSelectItem}
+          />
+        </Stack>
+        {/** 右半分(グラフ/メモ) */}
+        <Stack width="50%" justifyContent={"space-between"}>
+          {/** グラフ */}
+          <CircleGraph />
+          {/** メモ */}
+          <MemoList selectedItemTaskId={selectedItemTaskId} />
+        </Stack>
       </Stack>
-      {/** 右半分(グラフ/メモ) */}
-      <Stack width="50%" justifyContent={"space-between"}>
-        {/** グラフ */}
-        <CircleGraph />
-        {/** メモ */}
-        <MemoList selectedItemTaskId={selectedItemTaskId} />
-      </Stack>
-    </Stack>
+    </Suspense>
   );
 }
