--- my-app/src/app/work-log/task/page.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import Page from './page';
-
-const meta = {
-  component: Page,
-} satisfies Meta<typeof Page>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {}
-};
